updated roomba pathing logic
removed switch cases for the roomba pathing logic and shifted to if statements. roomba now paths to any dirty object to clean it and goes to its docking station if there is nothing to clean. roomba will always prioritize going after the player if the player is holding a dirty object. there are still a few bugs to flesh out.