functional ceiling fan in test level
ceiling fan is now fully functional within the current test level. Celing fan is related to the switch which can turn the fan on and off. the only way for the switch to turn on and off is by players jumping on the switch when it is turned on to turn it off and by jumping and headbutting the switch from the off position to turn the fan back on.