added in particles effects for throwing the coins out the window and started on clean in sink puzzle
I modified ShinyObject, by adding seralized fields and made it so when it the player scores it does the same particles effects as the one I have for the run into it coins.

The next thing I worked on is the sink, now this is extremly work in progress, and I havent been able to fully test it. So for this, I have made a sink prefab and a new script called SinkCleaner