fixing the Crowley sinking issue.
Found the source of the Crowley sinking issue. It is a float value that helps keep the player above ground. I added 0.1 to the value and now the sinking is less prevalent. We will have to continue to limit test it. If it proves to be too much of an issue and changing the value is not enough, the deeper problem most likely lies in the if statement below that checks if the player is sinking.